added some bots

+++ preprocess.py
@@ -20,7 +20,7 @@ def load_data(filename, seq_len, normalise_window):
     if normalise_window:
         result = normalise_windows(result)
 
-    result = np.array(result)
+    result = np.array(result, dtype='float32')
 
     row = round(0.9 * result.shape[0])
     train = result[:int(row), :]
